+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 * Throw an error is `boxes` elements in `bboxes_coords` are not integers.
+* Explicitly allow numpy array as arguments to `BBoxes.draw`
 
 ### Fixed